WIP of Gary Hudson trying to set some standards. See notes below.
Grid doubled in density so that can go back to regular 10 & 10 reference when not as many lines needed for placement
Experimented 2 x 2 x 2  (or 4  x 4 x 4 in double grid) for turnout standard spaces. This in turn means track spacing is 2 squares between . Font size for Gary block train text experimented with as 10. Is green (current the right choice or is something else better. Red ?)